# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Books API",
        "version": "0.1"
    },
    "paths": {
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get a list of books",
                "description": "Display a listing of the resource.",
                "operationId": "a9d13283dfab262e97af3c914d51b7a8",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Number of page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Search by Category",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "provider_id",
                        "in": "query",
                        "description": "Search by Provider",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Field to order by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at",
                            "enum": [
                                "title",
                                "author",
                                "created_at",
                                "updated_at"
                            ]
                        }
                    },
                    {
                        "name": "order_type",
                        "in": "query",
                        "description": "Order type (ascending or descending)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BookResource"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get a specific book by ID",
                "description": "Display the specified resource.",
                "operationId": "2fdd668eb661591758aff4c56ca9b66a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "The Book does not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BookResource": {
                "title": "Book Resource",
                "description": "Book resource representation",
                "properties": {
                    "id": {
                        "description": "ID of the book",
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "description": "Title of the book",
                        "type": "string"
                    },
                    "author": {
                        "description": "Author of the book",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the book",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Date and time when the book was created",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date and time when the book was last updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "category": {
                        "$ref": "#/components/schemas/CategoryResource"
                    },
                    "provider": {
                        "$ref": "#/components/schemas/ProviderResource"
                    }
                },
                "type": "object"
            },
            "CategoryResource": {
                "title": "Category Resource",
                "description": "Category resource representation",
                "properties": {
                    "id": {
                        "description": "ID of the category",
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "description": "Name of the category",
                        "type": "string"
                    },
                    "slug": {
                        "description": "Slug of the category",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ProviderResource": {
                "title": "Provider Resource",
                "description": "Provider resource representation",
                "properties": {
                    "id": {
                        "description": "ID of the provider",
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "description": "Name of the provider",
                        "type": "string"
                    },
                    "slug": {
                        "description": "Slug of the provider",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}